Clean Up/Bug Fixing
Cleaned up the code while also fixing a bug where the col_names would replace actual data past the first initiation. For example, if the matching session ID was in row 20, the previous code would replace rows 19 and 20 with the headers and data of the users personal file. Now this just replaces row 20 with the users survey answers.